mergge

+++ src/player.h
@@ -891,13 +891,13 @@ public:
 			client->sendIcons(getClientIcons());
 		}
 	}
-	void sendMagicEffect(uint16_t type) const
+	void sendMagicEffect(uint8_t type) const
 	{
 		if (client) {
 			client->sendMagicEffect(position, type);
 		}
 	}
-	void sendMagicEffect(const Position& pos, uint16_t type) const
+	void sendMagicEffect(const Position& pos, uint8_t type) const
 	{
 		if (client) {
 			client->sendMagicEffect(pos, type);
